+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `access` property added to `site`
 - `access` property added to `protocol`
 - `observations` property added to `site`
+- `observations` property added to `place`
 
 ### Changed